extract tabcontent, add validation to postReservation
