+++ client/src/component/account/LogoutBtn.tsx
@@ -1,5 +1,5 @@
-import { useAuth0 } from "@auth0/auth0-react";
 import React from "react";
+import { useAuth0 } from "@auth0/auth0-react";
 
 export default function LogoutBtn() {
   const { logout } = useAuth0();
