+++ client/src/component/account/accountReservation.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect } from "react";
-import { useAccountDetails } from "../../context/AccountDetailsProvider";
+import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
-import Auth0LoginRedirectBtn from "../auth0/Auth0LoginRedirectBtn";
 import { NavLink } from "react-router-dom";
+import { useAccountDetails } from "../../context/AccountDetailsProvider";
+import Auth0LoginRedirectBtn from "../auth0/Auth0LoginRedirectBtn";
 
 const listStyle = `
   border 
@@ -13,7 +13,7 @@ const listStyle = `
 export default function AccountReservation() {
   const { reservations } = useAccountDetails();
   const { isAuthenticated } = useAuth0();
- 
+
   return (
     <div>
       {isAuthenticated && reservations && (
